Move posts into subdirectory in case multiple posts are written in the same day; add images from cloudinary url
