--- src/scribblesPosts/2016-09-11.tsx
@@ -1,90 +1,0 @@
-import React from 'react';
-
-const post = {
-  date: '2016-09-11',
-  title: 'Duterte and the Bud Dajo Massacre',
-  markdown: `
-      # hi At the East Asia Summit last Friday, Philippines President Rodrigo Duterte took the
-      opportunity to shock an audience of world leaders with the infamous black and white photograph
-      of the 1906 Bud Dajo Massacre, depicting American soldiers posing proudly in a volcano crater
-      littered with hundreds of corpses. In the wake of the president’s extrajudicial killings that
-      has seen the deaths of more than 2,000 drug offenders and pushers, contextualizing the
-      massacre at Bud Dajo is essential to understanding Duterte’s vitriolic comments about
-      outsiders criticizing him for violating human rights; Obama being called “the son of a whore,”
-      and the UN Secretary General San Ki-Moon, a “fool,” being just two examples of his uncouth
-      rhetoric that surpasses Trump in terms of impropriety. The Bud Dajo Massacre stands as one of
-      the bleaker moments in a war that often escapes American consciousness. This is the
-      Philippine-American War (1899-1902), which began at the tail end of the Spanish-American War
-      (1898). After Spain was defeated in the Battle of Manila Bay, Spain made her exit from the
-      Eastern Hemisphere. Filipino nationalists seized this opportunity to declare their country an
-      independent and sovereign state. However, the United States, eager to secure an economic
-      foothold in Asia, as well as wet her feet in the colonial game played by all significant
-      European powers, naturally assumed that the defeat of Spain meant the seamless transfer of
-      authority into her hands. And so began the Philippine-American War. This war on the Filipino
-      archipelago can best be described as a proto-Vietnam. With military superiority, the Americans
-      were trying to subjugate a hostile country whose geography constantly thwarted an easy
-      victory. The Filipinos utilized the hot, humid, Southeast Asian jungle to conduct guerrilla
-      style warfare, miring the Americans into a grueling conflict. Often frustrated by the
-      effectiveness of their enemies’ strategies, the Americans then began to muddy their reputation
-      as being morally superior to Europeans in terms of war conduct (see the LIFE magazine cover
-      above), by employing unethical practices which stirred the ire of Americans at home, including
-      notables like Mark Twain. For one, the American soldiers used concentration camps,
-      euphemistically titled “protected zones,” to intern the native peoples, which one commandment
-      referred to as “suburbs of Hell.” They also adopted a scorched earth policy that called for
-      the razing of entire villages. Moreover, foreshadowing Guantanamo’s waterboarding, the
-      Americans utilized the ironically titled “water cure,” a method of interrogation involving the
-      forced consumption of excessive amounts of water, which was then expelled either naturally
-      through vomiting, or by force. Using these methods, coupled with modern technology, the
-      Americans eventually forced the Filipino revolutionaries to surrender. However, while this
-      officially ended the war and made the Philippines a colony of the United States, fighting
-      continued to happen, especially in the southern region of the country where Duterte himself is
-      from, the area of Mindanao. Mindanao is a unique region of the Philippines. Historically, the
-      peoples of Mindanao and the southern islands had a vastly differing ethnic, religious, and
-      linguistic identity from those living on the main island of Luzon. For one, most of these
-      Filipinos were identified as Moros, Spanish for the word Moors, i.e. the Muslim population of
-      the Iberian Peninsula expelled in the Spanish Reconquista. While Mindanao was an economically
-      viable area due to its proximity to important South Pacific trade routes, the Moros were
-      reputed for lawlessness, piracy, slavery, and tribalism. In addition, they ignored any form of
-      national law, instead basing their legal system on Islamic sharia. Even the Spanish had failed
-      to fully pacify Moroland, going so far as to transplant Christianized Filipinos from the north
-      into the area, a tactic that did nothing more than create further antipathy towards all
-      outsiders. The American strategy to handling the Moros involved making an agreement with the
-      local datus, or chieftains, typically via some form of a bribe. America promised to stay out
-      of the affairs of the Moros, so long as the Moros paid taxes and maintained stability. But
-      unbeknown to the Americans, the Moros were never a centralized people. Ratifying a treaty with
-      one datu, therefore, did not promise that the Moros under the authority of a different datu
-      would not offer resistance. And so, predictably, some of the Moros began to rebel, seeing the
-      American imperialist as no different than the Spaniard. With certain slopes at steep 60 degree
-      angles, and with an eleven mile circumference, the crater of the extinct Bud Dajo volcano was
-      the perfect hideout for those recalcitrants that refused to concede to the Americans. It was
-      self sufficient, containing irrigated water pools and patches along the outer rim to grow
-      crops. It was also very well defended, with multiple defensive fortifications built along the
-      crest. Aspiring to a famed military reputation, the local American governor felt compelled to
-      subjugate this stronghold of criminals and their families. And so, from March 7-9 1906, a
-      force of 800 Americans launched a difficult uphill struggle against this Moro outpost. The
-      Moros offered a fierce resistance, but once the Americans were able to get a machine gun
-      safely situated at the rim of the volcano, 600 men, women, and children were killed en masse,
-      described by one veteran as, “dominoes falling.” A photographer immortalized their victory
-      with the same photograph Duterte displayed at the Far East Asia Summit. In the postcolonial
-      world that’s haunted with issues of identity (think of the agonizing paradox that historically
-      colonized people are often forced to articulate their struggle for identity through the
-      language of their colonizers), formerly colonized countries are weary when others, especially
-      their former colonizer, try to impose a moral reproach. Hence why Duterte reminds the world of
-      the war that resulted in the death of more than 200,000 dead Filipinos, a war which America
-      has yet to issue a formal apology. Naturally, when Obama denounces his policies, Duterte
-      revives America’s skeletons-in-her-closet, as a way to kick her off her moral high ground. Be
-      that as it may, that doesn’t excuse Duterte from criticism by America or the United Nations.
-      And while I’ll concede ignorance to to the reality of drug trafficking and addiction on
-      Manila’s streets, I will acknowledge that the proposed quick-fix solution of having 100,000
-      bodies floating in Manila Bay by the end of Duterte first six months in office won’t end the
-      problem of drugs, just as building a wall won’t prevent people south of the border from coming
-      into the United States, or Syrian refugees flooding into Europe. Instead what Duterte needs to
-      ask is: what are the motivations that push people into the world of drugs? A sense of feeling
-      economically disenfranchised? A sense of futility towards fighting an extremely corrupt
-      government? A lack of general welfare? Or is it something within the culture? These are the
-      questions that Duterte needs to ask himself and the people of the Philippines. Anything less
-      would be irresponsible.
-  `
-};
-
-export default post;
