--- src/scribblesPosts/2017-04-20.tsx
@@ -1,60 +1,0 @@
-const post = {
-  date: '2017-04-20',
-  title: 'Inside the World of Tulpas, aka Imaginary Friends for Adults',
-  markdown: `
-      Direct experience is an evasion, or hiding place, for those without any imagination. –
-      Fernando Pessoa Where’s your will to be weird? – Jim Morrison Just as it sounds, the word,
-      “tulpa” is not native to the English language. It originates from a Buddhist practice of
-      fabricating a spiritual being solely through extreme focus and thought. But with the advent of
-      the Internet, tulpas became inspired with contemporary meaning, thanks to a very surprising
-      source, the Brony movement. Fanboys of My Little Pony: Friendship is Magic reappropriated the
-      term tulpa to signify an imaginary companion, also created solely by willpower–albeit in their
-      context, typically ponies from the TV show. Since the early 2010s, tulpas began to include any
-      being, in whatever form, from whatever source, created by a “tulpamancer” Home to many
-      underground communities, 4chan was one of the first crossroads where tulpamancers could meet
-      and share their ideas. Some members have since migrated to forums, as well as the subreddit
-      /r/tulpas, where you’ll discover numerous guides on how to create and nourish your tulpa, as
-      well as countless Q&As addressing every single angle. Here’s my quick rundown of tulpas and
-      tulpamancers: Tulpas are created to assist in memory recall, to help get rid of negative
-      feelings, but most importantly, for companionship. This is purportedly a lifelong commitment,
-      therefore a high level of maturity is believed to be a requisite (one guide explicitly says do
-      not create a tulpa if you’re young). The tulpa creation process can potentially take months to
-      years. To create a tulpa, it helps to first imagine a “wonderland,” or an imaginary
-      environment for your tulpa to inhabit. Tulpas are brought to life via “forcing,” either
-      “actively,” through a practice similar to meditation, or “passively,” where you train yourself
-      to direct all of your daily thoughts to the tulpa. It is unhealthy to assume that your tulpa
-      will take on the form you originally intend. Tulpas are believed to be organic, individual,
-      stand-alone entities, therefore some “deviation” is expected, just as you can’t expect your
-      friends to be develop exactly how you want them to be. You can create multiple tulpas,
-      although creating more than five is discouraged, just as spreading yourself too thin amongst
-      too many best friends means you can’t give each one the same concentrated attention.
-      “Switching” is a much more mature concept where you can literally exchange places with your
-      tulpas, i.e. inhabit your “wonderland” while your tulpa takes control of your body and speech.
-      Tulpamancers are an inherently shy group. Anybody who’s been on the receiving end of
-      harassment as a kid knows the source of this defensiveness. Conditioning teaches you that
-      opening your mouth leads to ridicule, so thoughts worth having are not thoughts worth sharing.
-      Therefore, whenever someone–journalist, academic, psychologist, or just simply a naturally
-      inquisitive person–requests an interview with a tulpamancer, the top comments are always
-      warnings by the veterans, sometimes outright demanding that no one participate, often citing
-      the 2014 Vice article which they believe spun the tulpamancer community as a group of
-      basement-dwelling perverts, clutching their Twilight Sprinkle plushies. However, as
-      tulpamancers do strive to gain a greater acceptance–emboldened as they are by online
-      communities–some have begun to ally themselves with the Hearing Voices Movement, advocacy
-      groups trying to destigmatize the Western diagnosis of disorders like schizophrenia, arguing
-      that auditory hallucinations are actually a positive, healthy attribute, as evidenced by it’s
-      greater appreciation amongst some non-Western, less globally-involved cultures. The obvious
-      foreseeable danger then becomes a muddied distinction between private hobby and genuine mental
-      illness, an effect that could prove highly calamitous to some. But knee-jerk reactions aside,
-      overall, are tulpas really that outlandish? How different is it than any other belief system
-      that peoples negative space with all kinds of fantastical creatures, gods and devils, angels
-      and demons, spirits and ghosts? Don’t these projections all originate from the same place, to
-      not feel so abysmally lonely? From skimming multiple threads on /r/tulpas, you soon discover
-      that the question of whether or not tulpas are real is irrelevant, especially to these
-      tulpamancers. Like any religious or spiritual belief that provides a greater sense of
-      authentic meaning and purpose: if it helps, it helps. In the sage words of George Costanza,
-      it’s not a lie…if you believe it. Reddit threads, further reading for the curious:
-    </Markdown>
-  `,
-};
-
-export default post;
