+++ src/scribblesPosts/posts.ts
@@ -1,4 +1,4 @@
-import post1 from './2016-09-11';
-import post2 from './2017-04-20';
+import post1 from './2016-09-11/2016-09-11';
+import post2 from './2017-04-20/2017-04-20';
 
 export default [post1, post2];
